anomaly-detector/usecases/nyc_taxi/requirements.txt
kafka-python==2.2.10
pandas
pyarrow